refactor(library/Attempt): moves "Failure/definition.ts" to "Failure/dueTo.ts"
